refactor(common): organize data dto entity packages (#33)
* refactor(common): organize data dto entity packages

* refactor(manager): split runspec subpackages
