+++ internal/common/data/run/spec/spec.go
@@ -2,7 +2,7 @@ package spec
 
 import (
 	"github.com/google/uuid"
-	"github.com/seedspirit/nano-backend.ai/internal/common/run/preset"
+	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/preset"
 )
 
 // Spec is the immutable, finalized input used to create a Run.
