+++ internal/manager/runspec/processor/candidate.go
@@ -1,8 +1,8 @@
-package runspec
+package processor
 
 import (
-	"github.com/seedspirit/nano-backend.ai/internal/common/run/draft"
-	"github.com/seedspirit/nano-backend.ai/internal/common/run/preset"
+	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/draft"
+	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/preset"
 )
 
 // Candidate is the validation target created from user draft and resolved presets.
