+++ internal/manager/runspec/processor/errors.go
@@ -1,4 +1,4 @@
-package runspec
+package processor
 
 import "strings"
