+++ internal/manager/runspec/processor/processor.go
@@ -1,12 +1,12 @@
-package runspec
+package processor
 
 import (
 	"context"
 
 	"github.com/google/uuid"
-	"github.com/seedspirit/nano-backend.ai/internal/common/run/draft"
-	"github.com/seedspirit/nano-backend.ai/internal/common/run/preset"
-	"github.com/seedspirit/nano-backend.ai/internal/common/run/spec"
+	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/draft"
+	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/preset"
+	"github.com/seedspirit/nano-backend.ai/internal/common/data/run/spec"
 	"github.com/seedspirit/nano-backend.ai/internal/manager/errordef"
 )
